8_5/requirements.txt
asgiref==3.8.1
Django==5.0.9
django-cleanup==9.0.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
